requirements.txt
customtkinter>=5.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.15.0
python-pptx>=0.6.21
FlagEmbedding>=1.2.0
requests>=2.31.0